requirements.txt
flask==2.3.2
python-dotenv==1.0.0
azure-search-documents==11.4.0
azure-core==1.28.0
gunicorn==21.2.0
